#138 logs tailing - moving tail to instance plugin
