+++ src/main/kotlin/com/cognifide/gradle/aem/instance/tail/Parser.kt
@@ -1,4 +1,4 @@
-package com.cognifide.gradle.aem.tooling.tail
+package com.cognifide.gradle.aem.instance.tail
 
 import java.io.BufferedReader
